requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.3.2
Django==5.0.1
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.27.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pypng==0.20220715.0
python-telegram-bot==21.0.1
qrcode==7.4.2
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
yarl==1.9.4
